[refactor] TestWatcher 코드 위치 분리 및 로그 재작성
- RepeatTestWatcher(TestWatcher 상속 클래스)를 CalculatorTest에서 분리
- 단위테스트 시작, 종료 로그 재작성
