+++ src/test/java/com/naver/commerce_platform/junit/CalculatorTest.java
@@ -3,10 +3,7 @@ package com.naver.commerce_platform.junit;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestRule;
-import org.junit.rules.TestWatcher;
-import org.junit.runner.Description;
 import org.junit.runner.RunWith;
-import org.junit.runners.model.Statement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,38 +14,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class CalculatorTest {
 
     static final Logger logger =
-            LoggerFactory.getLogger(RepeatRunner.class);
+            LoggerFactory.getLogger(CalculatorTest.class);
 
     @Rule
-    public final TestRule watchman = new TestWatcher() {
-        @Override
-        public Statement apply(Statement base, Description description) {
-            return super.apply(base, description);
-        }
-
-        @Override
-        protected void succeeded(Description description) {
-            logger.info("{} unit test succeeded.", description.getMethodName());
-        }
-
-        @Override
-        protected void failed(Throwable e, Description description) {
-            logger.error("{} unit test failed with {}.", description.getMethodName(), e);
-        }
-
-        @Override
-        protected void starting(Description description) {
-            super.starting(description);
-            logger.info("Run {}...", description);
-        }
-
-        @Override
-        protected void finished(Description description) {
-            super.finished(description);
-            logger.info("End {}...", description);
-        }
-    };
-
+    public final TestRule watchman = new RepeatTestWatcher();
 
     @Test
     @Repeat(count = 5)
